3ada6dd589172a1245779730a4e142fb44d7f.

+++ src/caffe/util/im2col.cpp
@@ -46,9 +46,10 @@ inline void im2col_nd_core_cpu(const Dtype* data_input, const bool im2col,
     const int* kernel_shape, const int* pad, const int* stride,
     Dtype* data_output) {
   if (!im2col) {
-    int im_size = im_shape[0];
+    //int im_size = im_shape[0];
+    size_t im_size = im_shape[0];
     for (int i = 0; i < num_spatial_axes; ++i) {
-      im_size *= im_shape[1 + i];
+      im_size *= (size_t)im_shape[1 + i];
     }
     caffe_set(im_size, Dtype(0), data_output);
   }
@@ -71,17 +72,21 @@ inline void im2col_nd_core_cpu(const Dtype* data_input, const bool im2col,
     for (bool incremented = true; incremented; ) {
       // Loop over spatial axes in forward order to compute the indices in the
       // image and column, and whether the index lies in the padding.
-      int index_col = c_col;
-      int index_im = c_col / kernel_size;
+    //  int index_col = c_col;
+    //  int index_im = c_col / kernel_size;
+      size_t index_col = c_col;
+      size_t index_im = c_col / kernel_size;
       bool is_padding = false;
       for (int d_i = 0; d_i < num_spatial_axes; ++d_i) {
-        const int d = d_iter[d_i];
-        const int d_im = d * stride[d_i] - pad[d_i] + d_offset[d_i];
+        //const int d = d_iter[d_i];
+        //const int d_im = d * stride[d_i] - pad[d_i] + d_offset[d_i];
+        const size_t d = d_iter[d_i];
+        const size_t d_im = d * stride[d_i] - pad[d_i] + d_offset[d_i];
         is_padding |= d_im < 0 || d_im >= im_shape[d_i + 1];
-        index_col *= col_shape[d_i + 1];
-        index_col += d;
-        index_im *= im_shape[d_i + 1];
-        index_im += d_im;
+        index_col *= (size_t)col_shape[d_i + 1];
+        index_col += (size_t)d;
+        index_im *= (size_t)im_shape[d_i + 1];
+        index_im += (size_t)d_im;
       }
       if (im2col) {
         if (is_padding) {
